tests/wpt/tests/css/css-ui/parsing/user-find-valid.tentative.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 4: parsing user-find with valid values</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/3460">
<meta name="assert" content="user-find supports the grammar 'auto | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("user-find", "auto");
test_valid_value("user-find", "none");
</script>
</body>
</html>
